requirements/base.txt
asgiref==3.2.10
Django==3.0
django-mssql-backend==2.8.1
pyodbc==4.0.30
pytz==2020.1
sqlparse==0.3.1
django-crispy-forms==1.9.2
django-jquery==3.1.0
django-tempus-dominus==5.1.2.13
django-filter==2.4.0